backend/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
Django==3.2.16
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
iniconfig==2.0.0
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
PyJWT==2.6.0
pytest==7.2.1
pytest-django==4.5.2
pytz==2022.7.1
redis==4.4.2
sqlparse==0.4.3
